+++ examples/vectorIndexAnthropic.ts
@@ -1,20 +1,20 @@
-import fs from "fs/promises";
+import fs from "node:fs/promises";
+
 import {
   Anthropic,
   Document,
-  VectorStoreIndex,
   serviceContextFromDefaults,
+  VectorStoreIndex,
 } from "llamaindex";
 
 async function main() {
   // Load essay from abramov.txt in Node
-  const essay = await fs.readFile(
-    "node_modules/llamaindex/examples/abramov.txt",
-    "utf-8"
-  );
+  const path = "node_modules/llamaindex/examples/abramov.txt";
+
+  const essay = await fs.readFile(path, "utf-8");
 
   // Create Document object with essay
-  const document = new Document({ text: essay });
+  const document = new Document({ text: essay, id_: path });
 
   // Split text and create embeddings. Store them in a VectorStoreIndex
   const serviceContext = serviceContextFromDefaults({ llm: new Anthropic() });
@@ -25,7 +25,7 @@ async function main() {
   // Query the index
   const queryEngine = index.asQueryEngine();
   const response = await queryEngine.query(
-    "What did the author do in college?"
+    "What did the author do in college?",
   );
 
   // Output response